% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitch_auth.R
\name{Auth}
\alias{Auth}
\alias{twitch_auth}
\title{Authentication}
\usage{
twitch_auth()
}
\description{
Authentication
}
\references{
https://github.com/Freguglia/rTwitchAPI/blob/master/R/authentication.R
}
