% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clip.R
\name{get_clip}
\alias{get_clip}
\title{Get Clip
@return A list with clip information
@export
@references https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
@import dplyr
@import httr}
\usage{
get_clip(clip_id)
}
\description{
Get Clip
@return A list with clip information
@export
@references https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
@import dplyr
@import httr
}
