% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clip_data.R
\name{get_clip_data}
\alias{get_clip_data}
\title{Get Clip Data
@return A list with clip information
@references https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
@import dplyr
@import httr}
\usage{
get_clip_data(clip_id)
}
\description{
Get Clip Data
@return A list with clip information
@references https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
@import dplyr
@import httr
}
