% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vod_data.R
\name{get_vod_data}
\alias{get_vod_data}
\title{Get Vod Data
@return VOD INFO
@export
@references https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
@import dplyr
@import httr}
\usage{
get_vod_data(vod_id)
}
\description{
Get Vod Data
@return VOD INFO
@export
@references https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
@import dplyr
@import httr
}
