% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rchamp-package.R
\docType{package}
\name{Rchamp-package}
\alias{Rchamp}
\alias{Rchamp-package}
\title{Rchamp: Get chat data from twitch clips}
\description{
This package starts from setting up twitch auth, getting raw data from twitch clips and then parsing it into a useable dataframe for consumption.
}
\author{
\strong{Maintainer}: Michael Puerto \email{michael.puerto11@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
