% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clip_chat.R
\name{get_clip_chat}
\alias{get_clip_chat}
\title{Get Clip chat by using vod info}
\usage{
get_clip_chat(clip_id, json_output = FALSE)
}
\description{
@description Uses the clip slug to determine the clip offset and gather the chat from the clip by navitating to the VOD at the offset.
}
\details{
@param clip_id the clip slug. Can be found in the url of the twitch clip.

@return Messages, usernames as list. Message information as json.

@references https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
@references https://github.com/OgulcanCelik/twitch-clip-chat

@import dplyr
@import httr
@import httr
@import jsonlite
}
