% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clip_chat.R
\name{get_clip_chat}
\alias{get_clip_chat}
\title{getting clip chat
@return A list with clip information
@export
@references https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
@import dplyr
@import httr}
\usage{
get_clip_chat()
}
\description{
getting clip chat
@return A list with clip information
@export
@references https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
@import dplyr
@import httr
}
