% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vod_data.R
\name{get_vod_data}
\alias{get_vod_data}
\title{Get Vod Data}
\usage{
get_vod_data(vod_id)
}
\arguments{
\item{vod_id}{this vod id can be found after using the get_clip_data function.}
}
\value{
VOD INFO
}
\description{
Get Vod Data
}
\references{
https://github.com/Freguglia/rTwitchAPI/blob/master/R/get_clip.R
}
